+++ src/main/java/be/woutschoovaerts/mollie/data/order/OrderLineCategory.java
@@ -1,0 +1,15 @@
+package be.woutschoovaerts.mollie.data.order;
+
+import com.fasterxml.jackson.annotation.JsonValue;
+
+public enum OrderLineCategory {
+
+    MEAL,
+    ECO,
+    GIFT;
+
+    @JsonValue
+    public String getJsonValue() {
+        return this.name().toLowerCase();
+    }
+}