+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-## 3.3.0
+## 3.2.2
 - Allow a custom user agent string
 
 ## 3.2.0
+- Add support for category attribute in OrderLineRequest
 - Exposed the dashboard link in the OrderLink
+
 
 ## 3.1.0
 - Fixed camel case issue for 'metadata' in OrderRequest